Restructure GraphQL directory from data to graphQL and split into queries and schemas, closes #17
