+++ graphQL/queries/index.js
@@ -1,0 +1,1 @@
+export { default as swapiQueries } from './swapiQueries';
