+++ graphQL/schemas/index.js
@@ -1,0 +1,1 @@
+export { default as swapiSchema } from './swapiSchema';
